social_network/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.7,<5.0
django-filter>=2.4,<3.0
djongo>=1.3.6
